# openapi.yml
openapi: 3.0.3
info:
  title: Spoon配信サポートAPI
  description: Spoon配信者向けの配信活動支援ツールAPI
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.spoon-support.com/v1
    description: Production server
  - url: https://staging-api.spoon-support.com/v1
    description: Staging server
paths:
...
  /events:
    get:
      tags:
        - Events
      summary: イベント一覧取得
      security:
        - BearerAuth: []
      responses:
        '200':
          description: イベント一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsResponse'
    post:
      tags:
        - Events
      summary: イベント作成
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EventsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Event'
    Event:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, completed, cancelled]
        participantCount:
          type: integer
        templateConfig:
          type: object
        createdAt:
          type: string
          format: date-time
    CreateEventRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        templateConfig:
          type: object
      required:
        - name
        - startDate
        - endDate
...